requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
APScheduler==3.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0

